more qc

+++ modules/local/ace.nf
@@ -24,8 +24,10 @@ process ACE {
     def prefix = task.ext.prefix ?: "${meta.id}"
     """
     prefix=$prefix
-    mv $bam \$prefix.bam
-    mv $bai \$prefix.bai
+    if [ ! -e \$prefix.bam ]; then
+        mv $bam \$prefix.bam
+        mv $bai \$prefix.bai
+    fi
     ace.R .
 
     cat <<-END_VERSIONS > versions.yml
